% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustMC-package.R
\docType{package}
\name{ClustMC-package}
\alias{ClustMC}
\alias{ClustMC-package}
\title{ClustMC: Cluster-Based Multiple Comparisons}
\description{
Multiple comparison techniques are typically applied following an F test from an ANOVA to decide which means are significantly different from one another. As an alternative to traditional methods, cluster analysis can be performed to group the means of different treatments into non-overlapping clusters. Treatments in different groups are considered statistically different. Several approaches have been proposed, with varying clustering methods and cut-off criteria. This package implements cluster-based multiple comparisons tests and also provides a visual representation in the form of a dendrogram. Di Rienzo, J. A., Guzman, A. W., & Casanoves, F. (2002) <jstor.org/stable/1400690>. Bautista, M. G., Smith, D. W., & Steiner, R. L. (1997) \doi{10.2307/1400402}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SGS2000/ClustMC}
  \item \url{https://sgs2000.github.io/ClustMC/}
  \item Report bugs at \url{https://github.com/SGS2000/ClustMC/issues}
}

}
\author{
\strong{Maintainer}: Santiago Garcia Sanchez \email{santiagoesquel@gmail.com} [copyright holder]

}
\keyword{internal}
