% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgc.R
\name{dgc_test}
\alias{dgc_test}
\title{DGC Test}
\usage{
dgc_test(y, trt, alpha = 0.05, show_plot = T, console = T, abline_options, ...)
}
\arguments{
\item{y}{Either a model (created with \code{lm()} o \code{aov()}) or a numerical vector
with the values of the response variable for each unit.}

\item{trt}{If \code{y} is a model, it corresponds to the column indicating the
treatments. If \code{y} is a vector, it's a vector of the same length as \code{y}
with the treatments for each unit.}

\item{alpha}{Value equivalent to 0.05 or 0.01, corresponding to the
significance level of the test. The default value is 0.05.}

\item{show_plot}{Logical value indicating whether the constructed dendrogram
should be plotted or not.}

\item{console}{Logical value indicating whether the results should be printed
on the console or not.}

\item{abline_options}{List with optional arguments for the line in the
dendrogram.}

\item{...}{Optional arguments for the \code{plot()} function.}
}
\value{
A list with three \code{data.frame}:
\item{stats}{\code{data.frame} containing summary statistics by treatment.}
\item{groups}{data.frame indicating the group to which each treatment is
assigned.}
\item{parameters}{\code{data.frame} with the values used for the test.
\code{treatments} is the total number of treatments, \code{alpha} is the
significance level used, \code{c} is the cut-off criterion for the dendrogram
(the height of the horizontal line of the dendrogram), \code{q} is the 1-alpha
quantile of the distribution of Q (distance from the root node) under the
null hypothesis and \code{SEM} SEM is an estimate of the standard error of the
mean.}
}
\description{
Di Rienzo, Guzman and Casanoves test for multiple comparisons.
Implements a cluster-based method for identifying groups of nonhomogeneous
means.
}
\examples{
data("PlantGrowth")
# Using vectors -------------------------------------------------------
weights <- PlantGrowth$weight
treatments <- PlantGrowth$group
dgc_test(y = weights, trt = treatments, show_plot = FALSE)
# Using a model ------------------------------------------------------
model <- lm(weights ~ treatments)
dgc_test(y = model, trt = "treatments", show_plot = FALSE)
}
\references{
Di Rienzo, J.A., Guzmán, A.W., Casanoves, F. (2002): A multiple
comparisons method based on the distribution of the root node distance of a
binary tree. J. Agr.Biol. Environ. Stat. 7: 1-14.
}
\author{
Santiago Garcia Sanchez
}
