% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustMC-package.R
\docType{package}
\name{ClustMC-package}
\alias{ClustMC}
\alias{ClustMC-package}
\title{ClustMC: Cluster Based Multiple Comparisons}
\description{
Pruebas de comparaciones múltiples basadas en conglomerados.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SGS2000/ClustMC}
  \item Report bugs at \url{https://github.com/SGS2000/ClustMC/issues}
}

}
\author{
\strong{Maintainer}: Santiago Garcia Sanchez \email{santiagoesquel@gmail.com}

}
\keyword{internal}
