% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClustMC-package.R
\docType{package}
\name{ClustMC-package}
\alias{ClustMC}
\alias{ClustMC-package}
\title{ClustMC: Cluster-Based Multiple Comparisons}
\description{
Implements cluster-based multiple comparisons tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SGS2000/ClustMC}
  \item \url{https://sgs2000.github.io/ClustMC/}
  \item Report bugs at \url{https://github.com/SGS2000/ClustMC/issues}
}

}
\author{
\strong{Maintainer}: Santiago Garcia Sanchez \email{santiagoesquel@gmail.com} [copyright holder]

}
\keyword{internal}
