% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgc.R
\name{dgc_test}
\alias{dgc_test}
\title{Test DGC}
\usage{
dgc_test(y, trt, alpha = 0.05, show_plot = T, console = T, abline_options, ...)
}
\arguments{
\item{y}{O bien un modelo (creado con \code{lm()} o \code{aov()}) o bien un vector
numerico con los valores de la variable respuesta para cada unidad.}

\item{trt}{Si \code{y} es un modelo, corresponde a la columna que indica los
tratamientos. Si \code{y} es un vector, corresponde a un vector de la misma
longitud que \code{y} con los tratamientos para cada unidad.}

\item{alpha}{Valor equivalente a 0.05 o 0.01, correspondiente al nivel
de significación del test. Por defecto toma el valor 0.05.}

\item{show_plot}{Valor logico indicando si debe graficarse el dendrograma
construido o no.}

\item{console}{Valor logico indicando si deben imprimirse los resultados en
consola o no.}

\item{abline_options}{Lista con argumentos opcionales para la linea
del dendograma.}

\item{...}{Argumentos opcionales para la funcion \code{plot()}.}
}
\value{
Una lista con tres objetos \code{data.frame}:
\item{estadisticas}{\code{data.frame} con estadisticas resumen segun tratamiento.}
\item{grupos}{\code{data.frame} indicando el grupo al que es asignado cada
tratamiento.}
\item{parametros}{\code{data.frame} con los valores utilizados para el test.
\code{tratamientos} es el numero total de tratamientos, \code{alpha} es el nivel de
significación utilizado, \code{c} es el criterio de corte (indica la altura
de la linea horizontal del dendrograma), \code{q} es el cuantil 1-alpha de la
distribucion de Q (distancia del nodo raiz) bajo la hipotesis nula y
\code{SEM} es una estimacion del error estandar de la media.}
}
\description{
Prueba de Di Rienzo, Guzmán y Casanoves para comparaciones múltiples.
Implementa un método basado en clusters para identificar grupos de medias no
homogéneas.
}
\examples{
data("PlantGrowth")
# Utilizando vectores -------------------------------------------------------
pesos <- PlantGrowth$weight
tratamientos <- PlantGrowth$group
dgc_test(y = pesos, trt = tratamientos, show_plot = FALSE)
# Utilizando un modelo ------------------------------------------------------
modelo <- lm(pesos ~ tratamientos)
dgc_test(y = modelo, trt = "tratamientos", show_plot = FALSE)
}
\references{
Di Rienzo, J.A., Guzmán, A.W., Casanoves, F. (2002): A multiple
comparisons method based on the distribution of the root node distance of a
binary tree. J. Agr.Biol. Environ. Stat. 7: 1-14.
}
\author{
Santiago Garcia Sanchez
}
